requirements.txt
# Core dependencies
torch>=2.5.0
torchvision>=0.22.0
torchaudio>=2.5.0  # optional but included for completeness
open_clip_torch>=3.2.0
timm>=1.0.17

# Image & video handling
pillow>=10.0.0
opencv-python>=4.10.0
numpy>=1.26.0

# Video downloaders
yt-dlp>=2025.1.1
pytube>=15.0.0

# Utilities
tqdm>=4.67.0
requests>=2.32.0
safetensors>=0.4.0
regex>=2024.9.0
pyyaml>=6.0.1

# Optional extras for development
matplotlib>=3.9.0
jinja2>=3.1.0
sympy>=1.13.0
networkx>=3.2.0
